add alert functionality

+++ src/sample/Controller.java
@@ -1,21 +1,19 @@
 package sample;
 
-import com.sun.jndi.toolkit.url.UrlUtil;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import javafx.scene.control.Button;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
-import javafx.scene.control.TextField;
+import javafx.fxml.Initializable;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
-import jdk.nashorn.internal.objects.NativeError;
+import javafx.scene.input.MouseEvent;
+import javafx.scene.layout.Region;
 
-import java.awt.event.ActionEvent;
 import java.net.URL;
 import java.sql.*;
+import java.util.Properties;
 import java.util.ResourceBundle;
 
-public class Controller {
+public class Controller implements Initializable {
 
     public TextField id_text;
     public TextField year_text;
@@ -29,18 +27,16 @@ public class Controller {
     public Button create_btn;
     public Button update_btn;
     public Button delete_btn;
-
-    public void Initialize(URL url, ResourceBundle rb){
-        pushVehicles();
-    }
+    private Properties user;
+    private Properties password;
 
     public Connection getConnection(){
 
         Connection connect_object;
         try{
-            connect_object = DriverManager.getConnection("");
+            Class.forName("com.mysql.cj.jdbc.Driver");
+            connect_object = DriverManager.getConnection("jdbc:mysql://localhost:3306/CRUDAPPSCHEMA","root", "Divyan23" );
             return connect_object;
-
         }
         catch(Exception e){
             System.out.println("Error:" + e.getMessage());
@@ -72,8 +68,7 @@ public class Controller {
 
     public void pushVehicles(){
 
-        ObservableList<Vehicle> vehicles;
-        vehicles = getVehicles();
+        ObservableList<Vehicle> vehicles = getVehicles();
         id_column.setCellValueFactory(new PropertyValueFactory<>("id"));
         year_column.setCellValueFactory(new PropertyValueFactory<>("year"));
         make_column.setCellValueFactory(new PropertyValueFactory<>("make"));
@@ -84,23 +79,49 @@ public class Controller {
 
     public void createVehicle() throws SQLException {
 
+        if(id_text.getText().equals("") || year_text.getText().equals("") || make_text.getText().equals("") || model_text.getText().equals("")) {
+
+            Alert alert = new Alert(Alert.AlertType.ERROR, "Please fill all text fields!", ButtonType.OK);
+            alert.getDialogPane().setMinHeight(Region.USE_PREF_SIZE);
+            alert.show();
+
+        }
+        else{
         String sql_query = "INSERT INTO vehicles VALUES(" + id_text.getText() + "," + year_text.getText() + ",'" + make_text.getText() + "','" + model_text.getText() + "')";
         establishSQLConnection(sql_query);
         pushVehicles();
+        }
     }
 
     public void updateVehicle() throws SQLException {
 
-        String sql_query = "UPDATE vehicles SET year = " + year_text.getText() + ",make = '" + make_text.getText() + "', model = '" + model_text.getText() + "' WHERE id = " + id_text.getText() + "";
-        establishSQLConnection(sql_query);
-        pushVehicles();
+        if(id_text.getText().equals("") || year_text.getText().equals("") || make_text.getText().equals("") || model_text.getText().equals("")) {
 
+            Alert alert = new Alert(Alert.AlertType.ERROR, "Please fill all text fields!", ButtonType.OK);
+            alert.getDialogPane().setMinHeight(Region.USE_PREF_SIZE);
+            alert.show();
+
+        }
+        else {
+            String sql_query = "UPDATE vehicles SET year = " + year_text.getText() + ",make = '" + make_text.getText() + "', model = '" + model_text.getText() + "' WHERE id = " + id_text.getText() + "";
+            establishSQLConnection(sql_query);
+            pushVehicles();
+        }
     }
 
     private void deleteVehicle() throws SQLException {
-        String sql_query = "DELETE FROM vehicles WHERE id = " + id_text.getText() + "";
-        establishSQLConnection(sql_query);
-        pushVehicles();
+
+        if(id_text.getText().equals("") || year_text.getText().equals("") || make_text.getText().equals("") || model_text.getText().equals("")){
+
+            Alert alert = new Alert(Alert.AlertType.ERROR, "Please select a row in the table or add an ID in the text field to delete!", ButtonType.OK);
+            alert.getDialogPane().setMinHeight(Region.USE_PREF_SIZE);
+            alert.show();
+        }
+        else{
+            String sql_query = "DELETE FROM vehicles WHERE id = " + id_text.getText() + "";
+            establishSQLConnection(sql_query);
+            pushVehicles();
+        }
 
     }
 
@@ -109,7 +130,7 @@ public class Controller {
         Connection connect = getConnection();
 
         try(Statement statement = connect.createStatement()){
-            statement.executeQuery(sql_query);
+            statement.executeUpdate(sql_query);
         }
         catch (Exception e){
             System.out.println("Error:" + e.getMessage());
@@ -128,5 +149,19 @@ public class Controller {
         else if(actionEvent.getSource() == delete_btn){
             deleteVehicle();
         }
+    }
+
+    public void mouseClicked(MouseEvent mouseEvent) {
+
+        Vehicle vehicle = (Vehicle) main_table.getSelectionModel().getSelectedItem();
+        id_text.setText(String.valueOf(vehicle.getId()));
+        year_text.setText(String.valueOf(vehicle.getYear()));
+        make_text.setText(vehicle.getMake());
+        model_text.setText(vehicle.getModel());
+    }
+
+    @Override
+    public void initialize(URL location, ResourceBundle resources) {
+        pushVehicles();
     }
 }
